refactor(server): extract batch path helpers
Move SUPPORTED_EXTENSIONS, collect_audio_files, output_path_for,
and parse_formats into batch/paths.rs. Public batch::* paths
are unchanged.
